backend/requirements.txt
flask==3.1.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pyjwt==2.9.0
bcrypt==4.2.0
python-dotenv==1.0.1
